Update Python version & add switch equivalent
Python 3.10 introduced the match statement, which is capable of rather more
than shown here.  Python 3.10 is the current release matched by the Python base
docker image, so seems a reasonable baseline for others to be able to check
against.

For the comment, I chose to hint without side-tracking.

Also fix a typo in the bash switch statement, spotted while using as a
cross-reference.